infrastructure/tests/requirements.txt
pytest==7.4.3
aws-cdk-lib==2.162.1
constructs>=10.0.0,<11.0.0
